+++ Ramensoup/src/Ramensoup/Renderer/IndexBuffer.h
@@ -15,8 +15,8 @@ namespace Ramensoup
 
 		virtual void Bind() const = 0;
 		virtual void Unbind() const = 0;
-		virtual uint32_t GetCount() const = 0;
+		virtual size_t GetCount() const = 0;
 
-		static const std::shared_ptr<IndexBuffer> Create(const uint32_t* indicies, uint32_t count);
+		static const std::shared_ptr<IndexBuffer> Create(const uint32_t* indicies, size_t count);
 	};
 }
