+++ Ramensoup/src/Ramensoup/Renderer/Renderer.h
@@ -22,7 +22,7 @@ namespace Ramensoup
 		static void Submit(const std::shared_ptr<Mesh>& mesh, const std::shared_ptr<Material>& material, const glm::mat4& transform = glm::mat4(1.0f));
 		static void EndScene();
 
-		static void DrawIndexed(const std::shared_ptr<VertexBuffer>& vertexBuffer, const std::shared_ptr<IndexBuffer>& indexBuffer, uint32_t indexCount = 0);
+		static void DrawIndexed(const std::shared_ptr<VertexBuffer>& vertexBuffer, const std::shared_ptr<IndexBuffer>& indexBuffer, size_t indexCount = 0);
 
 
 		inline static void SetViewport(uint32_t x, uint32_t y, uint32_t width, uint32_t height)
@@ -43,7 +43,7 @@ namespace Ramensoup
 		struct Statistics
 		{
 			uint32_t DrawCallCount = 0;
-			uint32_t TotalIndexCount = 0;
+			size_t TotalIndexCount = 0;
 		};
 		inline static const Statistics& GetStatistics() { return s_Statistics; }
